% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggData.R
\name{aggData}
\alias{aggData}
\title{Aggregate IIASA Data Using Regional Mappings}
\usage{
aggData(
  iiasa_raw = gtapssp::iiasa_raw,
  corresp_reg = gtapssp::corresp_reg,
  aggTxtFile = NULL,
  group_cols = c("model", "scenario", "reg_gtap_code", "variable", "unit")
)
}
\arguments{
\item{iiasa_raw}{A list containing IIASA raw data, typically the output of the \code{updateData} function.
Default is \code{gtapssp::iiasa_raw}.}

\item{corresp_reg}{A data frame containing the correspondence table between regions and their codes.
Default is \code{gtapssp::corresp_reg}.}

\item{aggTxtFile}{Character. The path to a Gempack-style text file containing additional
regional mappings. If \code{NULL}, this step is skipped. Default is \code{NULL}.}

\item{group_cols}{Character vector. The column names to use for grouping the aggregated data.
Default is \code{c("model", "scenario", "reg_gtap_code", "variable", "unit")}.}
}
\value{
A data frame containing the aggregated IIASA data grouped by the specified columns
and years. The aggregation sums up the \code{value} column for each group.
}
\description{
This function aggregates IIASA data based on a specified regional correspondence table.
Optionally, it can use additional regional mappings from a Gempack-style text file to
refine the correspondence table. The aggregated data is grouped by specified columns and years.
}
\examples{
\dontrun{
# Example with an additional mapping file
agg_data <- aggData(
  iiasa_raw = gtapssp::iiasa_raw,
  corresp_reg = gtapssp::corresp_reg,
  aggTxtFile = "path/to/aggFile.agg"
)

# Example without an additional mapping file
agg_data <- aggData(
  iiasa_raw = gtapssp::iiasa_raw,
  corresp_reg = gtapssp::corresp_reg
)
}

}
\seealso{
\link[gtapssp]{updateData}, \link[dplyr]{group_by}, \link[tidyr]{drop_na}
}
