% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{corresp_reg_pre}
\alias{corresp_reg_pre}
\title{Regional Correspondence (Predefined)}
\format{
A data frame with 5 variables:
\describe{
\item{\code{reg_gtap_number}}{Numeric identifier for the GTAP region.}
\item{\code{reg_iso3}}{ISO 3-letter code for the region.}
\item{\code{reg_gtap_code}}{GTAP-specific code for the region.}
\item{\code{reg_gtap_name}}{GTAP-specific name for the region.}
\item{\code{country_gtap_name}}{Country name associated with the GTAP region.}
}
}
\source{
Internal package data.
}
\usage{
corresp_reg_pre
}
\description{
A dataset mapping predefined regions to their corresponding metadata.
}
\keyword{datasets}
