% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvtorda.R
\name{csvtorda}
\alias{csvtorda}
\title{Save CSV Files as RDA Objects}
\usage{
csvtorda(data_folder = "data/")
}
\arguments{
\item{data_folder}{A character string specifying the path to the folder containing \code{.csv} files. Defaults to "data/".}
}
\value{
NULL. The function performs file operations and saves \code{.rda} files but does not return any value.
}
\description{
This function reads all \code{.csv} files in a specified folder, converts each into
a data frame, and saves them as \code{.rda} files in the same folder. The name of
each saved object matches the name of its source \code{.csv} file (without the extension).
}
\examples{
\dontrun{
  csvtorda("data/") # Save all .csv files in the "data/" folder as .rda
}
}
