% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{genderDict}
\alias{genderDict}
\title{Gender Dictionary}
\format{
A data frame with 2 variables:
\describe{
\item{\code{gender_code}}{Code representing gender (e.g., "MALE" for male, "FEML" for female).}
\item{\code{gender}}{Description of the gender (e.g., "Male", "Female").}
}
}
\source{
Internal package data.
}
\usage{
genderDict
}
\description{
A dataset mapping gender codes to their respective descriptions.
}
\keyword{datasets}
