% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\title{Calculate Growth Rate by Year for Specified Groups}
\usage{
growth_rate(
  data,
  group_cols = c("model", "scenario", "reg_gtap_code", "variable", "unit"),
  year_col = "year",
  value_col = "value",
  growth_rate_col = "growth_rate",
  na_replace = 0
)
}
\arguments{
\item{data}{A data frame containing the input data for which growth rates are to be calculated.}

\item{group_cols}{A character vector specifying the columns to group by before calculating the growth rate.
Default is \code{c("model", "scenario", "reg_gtap_code", "variable", "unit")}.}

\item{year_col}{A string specifying the column representing the year. Default is \code{"year"}.}

\item{value_col}{A string specifying the column containing the values for which growth rates are calculated.
Default is \code{"value"}.}

\item{growth_rate_col}{A string specifying the name of the new column where the calculated growth rates will be stored.
Default is \code{"growth_rate"}.}

\item{na_replace}{A numeric value used to replace missing growth rates (\code{NA}). Default is \code{0}.}
}
\value{
A data frame containing the original data with an additional column, representing
the calculated growth rates as percentage changes.
}
\description{
This function calculates the growth rate of a specified value column grouped by specified columns
in the data. The growth rate is computed as the percentage change between consecutive years
within each group. Missing growth rates (\code{NA}) are replaced with a user-defined value.
}
\examples{
\dontrun{
# Example usage
growth_rates <- growth_rate(
  data = my_data,
  group_cols = c("model", "scenario", "reg_gtap_code", "variable", "unit"),
  year_col = "year",
  value_col = "value",
  growth_rate_col = "annual_growth_rate"
)
}

}
\seealso{
\link[dplyr]{mutate}, \link[dplyr]{group_by}, \link[dplyr]{arrange}
}
