% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{iiasa_raw}
\alias{iiasa_raw}
\title{IIASA Raw Data}
\format{
A list containing multiple components:
\describe{
\item{\code{index}}{A pointer or metadata structure (if applicable).}
\item{\code{data}}{A data frame with projections and the following variables:
\describe{
\item{\code{model}}{The name of the model used for projections (e.g., "IIASA GDP 2023").}
\item{\code{scenario}}{The scenario associated with the data (e.g., "SSP2").}
\item{\code{region}}{The region or country associated with the data (e.g., "Africa (R10)").}
\item{\code{variable}}{The variable being projected (e.g., "GDP|PPP").}
\item{\code{unit}}{The unit of measurement for the variable (e.g., "billion USD_2017/yr").}
\item{\code{year}}{The year of the projection.}
\item{\code{value}}{The projected value for the variable.}
}
}
\item{\code{meta}}{Metadata for the dataset, such as descriptions, references, and DOIs.}
\item{\code{region}}{A vector of all regions covered in the dataset.}
\item{\code{variable}}{A vector of all variables included in the projections.}
\item{\code{unit}}{A vector of all units used in the dataset.}
\item{\code{year}}{A vector of all years included in the projections.}
}
}
\source{
Data provided by IIASA. Refer to the original DOI: https://doi.org/10.5281/zenodo.10618931.
}
\usage{
iiasa_raw
}
\description{
A dataset containing raw projections of socioeconomic pathways (SSPs) provided by IIASA.
}
\keyword{datasets}
