% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{isoList}
\alias{isoList}
\title{ISO Country List}
\format{
A data frame with 2 variables:
\describe{
\item{\code{Region}}{The full name of the country or region.}
\item{\code{iso}}{ISO 3-letter country code.}
}
}
\source{
Internal package data.
}
\usage{
isoList
}
\description{
A dataset containing ISO country codes and their corresponding details.
}
\keyword{datasets}
