% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_from_zip.R
\name{read_csv_from_zip}
\alias{read_csv_from_zip}
\title{Combine CSV Files from ZIP Archives}
\usage{
read_csv_from_zip(zip_dir, zip_pattern, csv_pattern, combine_vertically = TRUE)
}
\arguments{
\item{zip_dir}{Directory containing the ZIP files.}

\item{zip_pattern}{Pattern to match the ZIP file names.}

\item{csv_pattern}{Pattern to match the CSV file names inside the ZIP archives.}

\item{combine_vertically}{Logical, if TRUE, the function combines all CSV files vertically into a single data frame.
If FALSE, returns a list of data frames, each named after its corresponding CSV file.}
}
\value{
Either a single combined data frame or a list of data frames, depending on the value of \code{combine_vertically}.
}
\description{
This function searches for ZIP files in a specified directory that match a given pattern.
It then extracts CSV files from these ZIP archives that match another specified pattern.
The function can either combine these CSV files vertically into a single data frame or return
them as separate data frames in a list. Each data frame in the list is named after its respective CSV file.
}
\examples{
# Assuming ZIP files are in 'path/to/zip/files'
combined_data <- combine_csv_from_zip(
  zip_dir = "path/to/zip/files",
  zip_pattern = "zip_file_pattern",
  csv_pattern = "csv_file_pattern",
  combine_vertically = TRUE
)

# To get a list of data frames instead of a combined one
data_list <- combine_csv_from_zip(
  zip_dir = "path/to/zip/files",
  zip_pattern = "zip_file_pattern",
  csv_pattern = "csv_file_pattern",
  combine_vertically = FALSE
)
}
