% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCorresp.R
\name{updateCorresp}
\alias{updateCorresp}
\title{Update Country Correspondence Data with IIASA Regions}
\usage{
updateCorresp(
  iiasa_raw = gtapssp::iiasa_raw,
  corresp_reg = gtapssp::corresp_reg,
  outputFile = NULL
)
}
\arguments{
\item{iiasa_raw}{A list containing IIASA raw data, typically the output of the \code{updateData} function.
Default is \code{gtapssp::iiasa_raw}.}

\item{corresp_reg}{A data frame containing the existing correspondence table, typically
\code{gtapssp::corresp_reg}. This table is merged with the processed IIASA region data.}

\item{outputFile}{Character. Optional. The file path where the updated correspondence data will be saved.
If \code{NULL}, the data will not be saved but returned instead. Default is \code{NULL}.}
}
\value{
If \code{outputFile} is \code{NULL}, returns a data frame with the updated correspondence data.
Otherwise, saves the data to the specified file and invisibly returns \code{NULL}.
}
\description{
This function processes and updates a correspondence table between country names
and their ISO3 codes using data from the IIASA raw dataset and an existing correspondence
table. It cleans the region names, standardizes country names, and ensures alignment
with the ISO3 standard. The updated correspondence data can either be saved to a file
or returned directly.
}
\details{
The function performs the following steps:
\itemize{
\item Extracts unique region names from the \code{iiasa_raw} dataset.
\item Cleans and standardizes country names, removing invalid entries such as "World" and entries with parentheses.
\item Handles specific name corrections, such as replacing "Micronesia" with "Federated States of Micronesia".
\item Maps country names to their ISO3 codes using the \code{countrycode} package.
\item Merges the cleaned IIASA region data with the existing correspondence table (\code{corresp_reg}).
\item Optionally saves the final updated correspondence data to the specified file.
}
}
\examples{
\dontrun{
# Update correspondence data and save it to a file
updateCorresp(
  iiasa_raw = gtapssp::iiasa_raw,
  corresp_reg = gtapssp::corresp_reg,
  outputFile = "data/corresp_iiasa_updated.rda"
)

# Update correspondence data and return it directly
updated_data <- updateCorresp(
  iiasa_raw = gtapssp::iiasa_raw,
  corresp_reg = gtapssp::corresp_reg
)
}

}
\seealso{
\link[countrycode]{countrycode}, \link[dplyr]{filter}, \link[dplyr]{mutate}, \link[dplyr]{right_join}
}
