% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateData.R
\name{updateData}
\alias{updateData}
\title{Update Data from IIASA Using \code{pyam}}
\usage{
updateData(
  pythonExePath = NULL,
  condaenv = TRUE,
  outputFile = "data/iiasa_raw.rda"
)
}
\arguments{
\item{pythonExePath}{Character. Path to the Python executable. If \code{NULL}, the function
will attempt to use a Conda environment specified by \code{condaenv}. Default is \code{NULL}.}

\item{condaenv}{Logical. If \code{TRUE}, the function will use a Conda environment named
"base" if \code{pythonExePath} is not provided. Default is \code{TRUE}.}

\item{outputFile}{Character. Path where the processed data will be saved as a \code{.rda} file.
Default is \code{"data/iiasa_raw.rda"}.}
}
\value{
A list containing the processed IIASA data:
\describe{
\item{index}{Indices of the IamDataFrame.}
\item{data}{Time series data.}
\item{meta}{Metadata.}
\item{region}{Regions.}
\item{variable}{Variables.}
\item{unit}{Units.}
\item{year}{Years.}
}
The data is also saved to the specified output file.
}
\description{
This function reads and processes data from the IIASA database using the Python
\code{pyam} package. It validates the Python environment, configures the appropriate
Python executable or Conda environment, and ensures the required Python module
is available. The resulting data is saved as a \code{.rda} file.
}
\details{
\strong{Note:} Downloading the database from IIASA may take several minutes depending on
the size of the database and the speed of the connection.

This function requires the \code{pyam} Python package to be installed in the specified
Python or Conda environment. Ensure that the Python environment is correctly set up
before calling the function. Downloading the IIASA database may take several minutes.
}
\examples{
\dontrun{
# Using a specific Python executable
updateData(pythonExePath = "C:/path/to/python.exe", outputFile = "data/iiasa_raw.rda")

# Using the default Conda environment
updateData(condaenv = TRUE, outputFile = "data/iiasa_raw.rda")
}

}
\seealso{
\link[reticulate]{use_python}, \link[reticulate]{use_condaenv}, \link[reticulate]{py_config}, \link[reticulate]{import}
}
